Переместил интерфейсы в папку Interfaces
