+++ MetricsAgent/DAL/Interfaces/IHddMetricsRepository.cs
@@ -1,0 +1,9 @@
+﻿using MetricsAgent.DAL.Interfaces;
+using MetricsAgent.DAL.Models;
+
+namespace MetricsAgent.DAL.Repositories
+{
+    public interface IHddMetricsRepository : IRepository<HddMetrics>
+    {
+    }
+}
