+++ MetricsAgent/DAL/Interfaces/IRamMetricsRepository.cs
@@ -1,0 +1,9 @@
+﻿using MetricsAgent.DAL.Interfaces;
+using MetricsAgent.DAL.Models;
+
+namespace MetricsAgent.DAL.Repositories
+{
+    public interface IRamMetricsRepository : IRepository<RamMetrics>
+    {
+    }
+}
